packages/former/pyproject.toml
[project]
name = "former"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click>=8.2.0",
    "colorama>=0.4.6",
    "huggingface-hub>=0.31.1",
    "jiwer>=3.1.0",
    "omegaconf>=2.3.0",
    "pandas>=2.2.3",
    "pillow>=11.2.1",
    "pydub>=0.25.1",
    "pyyaml>=6.0.2",
    "sentencepiece>=0.2.0",
    "textgrid>=1.6.1",
    "torch>=2.7.0",
    "torchaudio>=2.7.0",
]
